requirements.txt
tensorflow>=2.16.0
tf-keras>=2.15.0
tensorflow-probability[tf]>=0.16.0
earthengine-api>=0.1.317
pandas>=1.4.0
numpy>=1.21.0
xarray>=0.20.0
geopandas>=0.10.0
rasterio>=1.2.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0
nbformat>=5.7.0
pytest>=7.0.0
tqdm>=4.65.0